refactor: move period settings to parent container component
