+++ src/components/PeriodDimension/PeriodDimension.js
@@ -1,7 +1,17 @@
+import { useConfig, useDataQuery } from '@dhis2/app-runtime'
 import PropTypes from 'prop-types'
 import React from 'react'
 import { DIMENSION_ID_PERIOD } from '../../modules/predefinedDimensions.js'
 import PeriodTransfer from './PeriodTransfer.js'
+
+const userSettingsQuery = {
+    userSettings: {
+        resource: 'userSettings',
+        params: {
+            key: ['keyUiLocale'],
+        },
+    },
+}
 
 const PeriodDimension = ({
     onSelect,
@@ -9,6 +19,14 @@ const PeriodDimension = ({
     rightFooter,
     excludedPeriodTypes,
 }) => {
+    const { systemInfo } = useConfig()
+    const result = useDataQuery(userSettingsQuery)
+
+    const { calendar = 'gregory' } = systemInfo
+    const { data: { userSettings: { keyUiLocale: locale } = {} } = {} } = result
+
+    const periodsSettings = { calendar, locale }
+
     const selectPeriods = (periods) => {
         onSelect({
             dimensionId: DIMENSION_ID_PERIOD,
@@ -22,6 +40,7 @@ const PeriodDimension = ({
             rightFooter={rightFooter}
             dataTest={'period-dimension'}
             excludedPeriodTypes={excludedPeriodTypes}
+            periodsSettings={periodsSettings}
         />
     )
 }
