+++ src/components/PeriodDimension/__tests__/PeriodDimension.spec.js
@@ -2,6 +2,13 @@ import { shallow } from 'enzyme'
 import React from 'react'
 import PeriodDimension from '../PeriodDimension.js'
 
+jest.mock('@dhis2/app-runtime', () => ({
+    useConfig: () => ({ systemInfo: {} }),
+    useDataQuery: () => ({ data: { userSettings: { keyUiLocale: 'en' } } }),
+}))
+
+afterEach(jest.clearAllMocks)
+
 describe('The Period Dimension component', () => {
     let props
     let shallowPeriodDimension
