+++ src/components/PeriodDimension/__tests__/PeriodSelector.spec.js
@@ -2,13 +2,6 @@ import { shallow } from 'enzyme'
 import React from 'react'
 import PeriodTransfer from '../PeriodTransfer.js'
 
-jest.mock('@dhis2/app-runtime', () => ({
-    useConfig: () => ({ systemInfo: {} }),
-    useDataQuery: () => ({ data: { userSettings: { keyUiLocale: 'en' } } }),
-}))
-
-afterEach(jest.clearAllMocks)
-
 describe('The Period Selector component', () => {
     let props
     let shallowPeriodTransfer
